слегка допилить find subject

+++ src/tree.cpp
@@ -21,10 +21,15 @@ void TreeCtor ( Tree_t* tree ) {
 
 /*=====================================================================================*/
 
-void TreeDtor ( Tree_t* tree ) {
+TreeErr_t TreeDtor ( Tree_t* tree ) {
 
     assert(tree);
+
+    if (!tree->root) return TreeErr_t::EMPTY_TREE_ACT_ERR;
     DeleteNode(tree->root);
+    free(tree->buffer);
+
+    _RET_OK_
 
 }
 
@@ -56,7 +61,9 @@ TreeErr_t DeleteNode ( TreeNode* node ) {
     if (node->right != nullptr)
         DeleteNode (node->right);
 
-    free(node->data);
+    if (node->is_alloc == 1)
+        free(node->data);
+
     free(node);
 
     _RET_OK_
@@ -121,6 +128,8 @@ TreeErr_t SaveToDisk ( Tree_t* tree, const char* disk_name ) {
     FILE* disk = fopen( disk_name, "wb" );
     if (disk == nullptr) return TreeErr_t::FILE_OPEN_ERR;
 
+    if (!tree->root) return TreeErr_t::EMPTY_TREE_ACT_ERR;
+
     WriteToDisk(tree->root, disk);
 
     fclose(disk);
@@ -136,12 +145,17 @@ void WriteToDisk ( TreeNode_t* node, FILE* disk ) {
     assert(node);
     assert(disk);
 
-    fprintf(disk, "(%s", node->data);
+    fprintf(disk, "(\"%s\"", node->data);
 
     if (node->left)
         WriteToDisk(node->left, disk);
+    else 
+        fprintf(disk, " nil"); 
+    
     if (node->right)
         WriteToDisk(node->right, disk);
+    else 
+        fprintf(disk, " nil");
         
     fprintf(disk, ")");
 
@@ -154,10 +168,28 @@ TreeErr_t ReadFromDisk (Tree_t* tree, const char* filename ) {
     assert(tree);
     assert(filename);
 
+    _OK_STAT_
+
     FILE* file = fopen(filename, "rb");
     if (!file) return TreeErr_t::FILE_OPEN_ERR;
+    long long byte_num = FileByteCount(filename);
 
+    tree->buffer = (char*)calloc(byte_num + _buff_byte_padding_, sizeof(tree->buffer[0]));
+    if (!tree->buffer) return TreeErr_t::MEM_ALLOC_ERR;
+    int r = fread(tree->buffer, sizeof(tree->buffer[0]), byte_num + _buff_byte_padding_, file);
+    printf("read_c: %d\n", r);
 
+    size_t pos = 0;
+
+    tree->root = ReadNode(tree->buffer, &pos, &status, &tree->cpcty);
+
+    if (status == TreeErr_t::READ_SYNTAX_ERR ) {
+        //printf("%d %s\n", pos, tree->buffer);
+        TreeDump(tree, status, &tree->buffer[pos]);
+        return status; 
+
+    } else if (status != TreeErr_t::TREE_OK )
+        return status;
 
     _RET_OK_
 
@@ -165,7 +197,7 @@ TreeErr_t ReadFromDisk (Tree_t* tree, const char* filename ) {
 
 /*=====================================================================================*/
 
-TreeNode_t* ReadNode ( char* buffer, size_t* pos, TreeErr_t* status ) {
+TreeNode_t* ReadNode ( char* buffer, size_t* pos, TreeErr_t* status, size_t* cpcty ) {
 
     if (*status != TreeErr_t::TREE_OK) return nullptr;
 
@@ -174,30 +206,45 @@ TreeNode_t* ReadNode ( char* buffer, size_t* pos, TreeErr_t* status ) {
         TreeNode_t* node = nullptr;
         *status = AllocNode(&node);
         if (*status != TreeErr_t::TREE_OK) return nullptr;
-        *pos++;
+        (*cpcty)++;
+        (*pos)++;
 
         node->is_alloc = 0;
         size_t len = 0;
+
         node->data = ReadData(&buffer[*pos], &len);
         if (!node->data) {
             *status = TreeErr_t::READ_DATA_ERR;
             return nullptr;
         }
-        *pos += len;
+        node->data_hash = djb2hash(node->data);
+        (*pos) += len;
 
-        node->left = ReadNode( buffer, pos, status);
-        node->right = ReadNode( buffer, pos, status);
-        pos++; //TODO: пропуски пробелов
+        //printf("%s\n", &buffer[*pos]);
+        SKIP_SPACE_
+        node->left = ReadNode( buffer, pos, status, cpcty);
+        SKIP_SPACE_
+        node->right = ReadNode( buffer, pos, status, cpcty);
+        SKIP_SPACE_
+        (*pos)++;
+        //printf(" %s\n", &buffer[*pos]);
+
+        if (node->left && node->right) {
+            node->left->parent = node;
+            node->right->parent = node;
+        }
         return node; 
 
-    } else if (strncmp(&buffer[*pos], _nil_, _nil_len_) == 0) {
+    } else if (buffer[*pos] == 'n', buffer[*pos+1] == 'i', buffer[*pos+2] == 'l') {
 
-        *pos+=_nil_len_;
+        (*pos)+=3;
+        SKIP_SPACE_
         *status = TreeErr_t::TREE_OK;
         return nullptr;
 
     } else {
 
+        //printf("%c\n", buffer[*pos]);
         *status = TreeErr_t::READ_SYNTAX_ERR;
         return nullptr;
 
@@ -207,8 +254,11 @@ TreeNode_t* ReadNode ( char* buffer, size_t* pos, TreeErr_t* status ) {
 
 /*=====================================================================================*/
 
-char* ReadData (char* ptr, size_t* len) {
+char* ReadData ( char* ptr, size_t* len ) {
 
-
+    sscanf( ptr, "\"%*[^\"]\"%n", len );
+    //printf("\n\n\n%c\n\n", *(ptr + *len-1));
+    *(ptr + *len - 1) = '\0';
+    return ptr+1;
 
 }
